Refactor: Organized each component in subfolder named same as the page the component is used in.
