+++ app/weather/page.tsx
@@ -8,10 +8,10 @@ import {
   WeatherForecast,
   WeatherForecastData,
 } from "@/public/types/Weather";
-import WeatherChip from "@/components/WeatherChip";
+import WeatherChip from "@/components/weather/WeatherChip";
 import Image from "next/image";
 
-import HourlyWeatherChip from "@/components/HourlyWeatherChip";
+import HourlyWeatherChip from "@/components/weather/HourlyWeatherChip";
 export default function Page() {
   const [location, setLocation] = useState<Location>({
     latitude: "",
