+++ components/weather/WeatherChip.tsx
